requirements.txt
# InvoiceFlow Python Dependencies
# Pinned versions for reproducible builds

# Core Django Framework
Django==5.2.9
django-environ==0.12.0
django-csp==4.0
django-ratelimit==4.1.0

# Database
psycopg2-binary==2.9.11

# Web Server
gunicorn==23.0.0
whitenoise==6.11.0

# PDF Generation
weasyprint==66.0
fonttools==4.60.1
cssselect2==0.8.0
tinycss2==1.5.0
tinyhtml5==2.0.0
pydyf==0.11.0
pyphen==0.17.2
webencodings==0.5.1

# Compression
brotli==1.2.0
zopfli==0.4.0

# Email
sendgrid==6.12.5

# Security & Authentication
cryptography==46.0.3
pyotp==2.9.0
qrcode==7.4.2

# Monitoring & Logging
sentry-sdk==1.45.1

# Image Processing
pillow==12.0.0

# REST API
djangorestframework==3.15.2
drf-spectacular==0.28.0

# Utilities
asgiref==3.11.0
cffi==2.0.0
pycparser==2.23
python-dateutil==2.9.0
PyYAML==6.0.3
sqlparse==0.5.3
requests==2.32.3
psutil==7.0.0

# Testing
factory-boy==3.3.1
pytest==8.4.0
pytest-cov==6.1.1
pytest-django==4.10.0
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
cryptography==46.0.3
cssselect2==0.8.0
Django==5.2.9
django-csp==4.0
django-environ==0.12.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
factory-boy==3.3.1
fonttools==4.60.1
gunicorn==23.0.0
pillow==12.0.0
psutil==7.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
pyotp==2.9.0
pyphen==0.17.2
pytest==8.4.0
pytest-cov==6.1.1
pytest-django==4.10.0
python-dateutil==2.9.0
PyYAML==6.0.3
qrcode==7.4.2
requests==2.32.3
sendgrid==6.12.5
sentry-sdk==1.45.1
sqlparse==0.5.3
tinycss2==1.5.0
tinyhtml5==2.0.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.0
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
cryptography==46.0.3
cssselect2==0.8.0
Django==5.2.9
django-csp==4.0
django-environ==0.12.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
factory-boy==3.3.1
fonttools==4.60.1
gunicorn==23.0.0
pillow==12.0.0
psutil==7.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.11.0
pyotp==2.9.0
pyphen==0.17.2
pytest==8.4.0
pytest-cov==6.1.1
pytest-django==4.10.0
python-dateutil==2.9.0
PyYAML==6.0.3
qrcode==7.4.2
requests==2.32.3
sendgrid==6.12.5
sentry-sdk==1.45.1
sqlparse==0.5.3
tinycss2==1.5.0
tinyhtml5==2.0.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.0
bandit
black
djlint
isort
pip-audit
ruff
safety
Django==5.2.9
